Removed the ability for the scanner to peek at the next input.
To make it easier to locally reason about scanning, the API is limited
so that no information can be gleaned about unconsumed input other than
"it is not something we want to consume". This way, behavior is entirely
determined by the actual content of a reader macro, not any of the
following content which wasn't consumed.